+++ plugin.video.bravo/README.txt
@@ -4,6 +4,13 @@ plugin.video.bravo
 Kodi Addon for Bravo TV website
 for Kodi Isengard and later
 
+icon.png sourced from public domain : 
+https://commons.wikimedia.org/wiki/File:Bravo_TV.svg
+
+fanart.jpg sourced from public domain:
+https://commons.wikimedia.org/wiki/File:Bravo_TV.svg (converted to .jpg)
+
+Version 3.0.2 website change & cleanup
 Version 3.0.1 separate scraping routines for future functions
 Version 2.0.3 cleanup for release
 Version 2.0.2 website change
