+++ plugin.video.bravo/addon.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <addon id="plugin.video.bravo"
        name="Bravo"
-       version="3.0.1"
+       version="3.0.2"
        provider-name="t1m">
   <requires>
     <import addon="xbmc.python" version="2.20.0"/>
@@ -21,6 +21,6 @@
     <forum>http://forum.kodi.tv/showthread.php?tid=224738</forum>
     <website>www.bravo.com</website>
     <email></email>
-    <source>https://github.com/learningit/plugin.video.bravo</source>
+    <source>https://github.com/learningit/Kodi-plugins-source</source>
   </extension>
 </addon>
