+++ plugin.video.bravo/resources/__init__.py
